refactor: expose library for visualizer
