+++ main.go
@@ -3,6 +3,8 @@ package main
 import (
 	"fmt"
 	"os"
+
+	"lemin/lem"
 )
 
 func main() {
@@ -10,12 +12,12 @@ func main() {
 		fmt.Println("Usage: lem-in <file>")
 		os.Exit(1)
 	}
-	graph, ants, lines, err := ParseFile(os.Args[1])
+	graph, ants, lines, err := lem.ParseFile(os.Args[1])
 	if err != nil {
 		fmt.Println(err.Error())
 		return
 	}
-	path, ok := BFS(graph, graph.Start, graph.End)
+	path, ok := lem.BFS(graph, graph.Start, graph.End)
 	if !ok {
 		fmt.Println("ERROR: invalid data format")
 		return
@@ -24,7 +26,7 @@ func main() {
 		fmt.Println(l)
 	}
 	fmt.Println()
-	moves := Simulate(path, ants, graph.Start, graph.End)
+	moves := lem.Simulate(path, ants, graph.Start, graph.End)
 	for _, m := range moves {
 		fmt.Println(m)
 	}
